QUEST: The Goblin Tailor
CORE: Add functions to calculate the RSE location and race of the week
SCRIPTS: Update treasure chest of Ordelle's Caves, Gusgen Mines and Maze
of Shakhrami to obtain the keyitem Magic Pattern if you're in the
correct location and race of the week
SCRIPTS: Update Guttrix lua to give the quest and all the pieces of the
RSE gear, only the first time the quest is completed. If you drop any
piece of the RSE gear, Guttrix offer again in your RSE race week.