Attempt to implement remaining RDM Spells and Abilities, all the way up to Lv99
Implemented Abilities:

Spontaneity - Makes the next spell have no cast time or recast time, as if the caster was under the effect of Chainspell. The implementation mirrors Chainspell, but removes itself after a single spell

Stymie - One Hour Special. Grants a magic accuracy bonus to the next enfeeble spell. From the available information, it seems that as long as the mob is not immune to the status effect, then it will land unresisted. The implementation attempts to mirror this by checking for Stymie + Immunity, and returning unresisted if not immune in the function to calculate  resistance to a status effect that is used by all enfeebling spells.

Implemented Spells:

Addle - Increases the targets casting time and lowers their magic accuracy. The exact formula is not known, but the best available information says that the raw value is ~30%, and is boosted to ~50% if Saboteur is active, so this implementation assumes that is the case.

Temper - Grants the target the effect of Multi Strikes, which gives them Double Attack.

Formula used from BGWiki:
-Enhancing Magic Skill < 360 : DA % = 5%
-Enhancing Magic Skill ≥ 360 : DA % = floor( Enhancing Magic Skill - 300 ) ÷ 10 )
-Cap: 20%

TODO: Spell animations, confirm ability animations.