CoP 7-1: Chains & Bonds - Fix Mismatched Cutscenes
The Ulmia/Prishe Cutscene is supposed to happen at the Walnut Door, while the Louverance Cutscene is supposed to happen at the gate to Aquaducts.

These were mixed up, leading to awkward scenarios where you click the walnut door and are suddenly at the gate to aquaducts in a cs, or click the gate to aquaducts and walk into a room.