ASA 3: That Which Curdles Blood
Note: While the recipes for all the enfeebling kits are implemented, the cs is only coded to take the enfeebling kit of poison. This is due to the implementation of 1-2 always rolling, from what i've seen, poison currently, so I don't know which option corresponds to which of the other kits.