Core: Fix dedication loss on logoff/crash
Removed MOD_DEDICATION and MOD_DEDICATION_CAP from
charutils::AddExpBonus
Implemented a SubPower for EFFECT_DEDICATION (replacing
MOD_DEDICATION_CAP) allowing the variable to be updated as EXP is
earned. Tested on test server and it appears to be working as intended.
Status effect persisted through server shutdown and logging off and back
on.