CORE: Add Support to Restrict Traits By Expansion
With this change you can restrict a trait from being enabled if the corresponding expansion is not enabled in settings.lua If NULL, the trait is enabled.

TO restrict the trait, use the abbreviated code XXX where ENABLE_XXX is the variable in settings.lua.

Initially, this is applied to the new SOA Max HP Boost Updates. It can be expanded to other traits later.

If this approach works out well we can expand it to tag new spells / abilities to their appropriate era as well.