December Update: NPC ID Updates
Npc List SQL: These updates were done by the new next-gen npc id job. This job is way more accurate than the previous version due to the changes we made in reformatting npc_list.sql. A handful of zones were detected as shifted. Since this is the first run of the new job, I tried to do some manual spot checking in each shifted zone, and from the npcs I spot checked, things were looking correct to me. With that said, it's still possible there are bugs. If anyone finds any, please tag me in the issue or change so that I can look at why the job did not update it correctly. I apologize in advance for any incorrect ids :p

Scripts/Transport/Elevators - These were done by a post processing script, which shifts any ids the npc list job shifted in any lua xcript or transport/elevators SQL. This is just done by doing an ordered by descending find/replace. What this means is, if the script was out of sync with the npc ids before this change, it will still have the wrong id and need to be fixed manually - otherwise it will have obtained the new id as according to the batch job. Any scripts that get manually corrected due to being out of sync should be automatically picked up next time its id shifts. Help in verifying these is appreciated - if we get them all in sync, this job should be able to keep them in sync.

Very interested in any bugs that may be found so that we can improve the jobs further. If anyone is interested in the jobs, check the darkstar-utils repo.