Automated Jeuno's yearly Christmas music changing.
I almost did the summer music for the zones that the events using it apply to before deciding that would be better off done in a global for that event, and called into those zones.

Jeuno can stay separate because the music is not so much part of the "Starlight Festival" or "Happy New Years" events as it is overlapping with them for most of December.