CORE: added confrontation effects
Targetting will be disallowed on players and mobs with a confrontation
status different from your own.  The targetting is done with getPower(),
keep that in mind when scripting any of the confrontation effects.  Should
fix BCNM mobs running out of their battlefield (as well as unwanted
visitors entering the BCNM via wallhacking)