Scripts: made future use of EFFECT_ADDLE more intuitive after seeing some folks trying to implement mobskills using it get unexpected results.
All I did was change it back to Startlightknight's original behavior of doing the subtractive values in the effect, rather than every script that calls it, while retaining the same addMod/delMod change that he was asked to do before. This means you won't have to set negative power/subPower values when adding the effect to have the modifier reduced by that amount.