Scripts/SQL: Enhances Rice Ball effect + Hachiman TP latent fix
I've added the bonus effects of each rice ball to each piece of gear as food-activated latents using food ID as the parameter. I removed the old system for the rice ball as it was granting the boost when the gear was not equipped - that was not retail behaviour. Fix for issue #1963.

I've also corrected the TP latent on the pieces of the Hachiman set that have it. It now checks for TP > 1000 instead of TP > 100.