Scripts: AF hands quest + BLM AF3 start requirements fix
For whatever reason, players were able to start AF hands quest without having the corresponding AF2 quest completed. I've changed the check to make sure AF2 is completed before allowing hands to start.

BLM AF3 quest could be started when AF2 quest is accepted. This is not right - AF2 needs to be completed before AF3 can begin, so I added a check for that as well.