Scripts/SQL: Martyr/Devotion/Shikikoyo target + message
Made it so players cannot use those three abilities on themselves - while they can't use it on themselves while solo, they can use it on themselves when in a party.
Set the message upon using the ability (recovers HP/MP, regains TP), and the message when a player tries to use the ability on themselves. This required a message to be added in status.lua, as well as capping the amount of HP/MP given at max HP/MP - current HP/MP, or TP given at 300 - current TP.