Scripts/SQL: Wamouracampa TP moves, Cannonball spell def mod.
Mob: Implemented Amber Scutum, Heat Barrier, Vitriolic Spray, Cannonball, and Heat Barrier. The last two have a form check.
Scripted the stance changing. They change after roughly 45 seconds.
Corrected the range of some of the mobskills, and enabled them for Brass Borer.
Added param offcratiomod to monstertpmoves.lua - if it's nil (which it will be for pretty much everything), it will default to the standard mob attack. Else it will use what's pass on. So for Cannonball, it passes mob defense instead.

Spell: Passed on param.offcratiomod as part of params - if it's nil, it will default to caster's attack. For Cannonball, it passes caster's defense instead.