Core/Scripts/SQL: Implement increased roll range mod.
To do this, I added a binding in the core for setRange and getRange for abilities. So in the scripts it's just ability:setRange(ability:getRange() + mod).
Added a modifier that would be used for the increased range, and added it to Luzaf's Ring.
Modified all roll scripts + Double-Up so it will set the range to 8 (retail value before any mods) + the mod value.