SQL/Core/Scripts: MobMod for scripted 2hrs
This modifier will allow mobs to have 2hrs in their particular moveset in mob_skill.sql without being able to use it as a random tp move. If the mod is set to 1, each 2hr script will return 1 in the onMonSkillCheck function. This will allow all mobs with 2hrs that are unscripted to use them as normal.

When an ability is forced to be used by the script, it bypasses the onMobSkillCheck function, and just uses the skill. This is why I didn't need to add an additional check. Also, since the skill parameter is used to pass along the duration for some of these scripted 2hrs, I modified the check to use the param value if the mobmod is set and the param value is not 0. Otherwise, it defaults to the standard duration.

I looked through mob_skill.sql to see what mob families had 2hrs in their skillset, and determined whether they needed the new mobmod, or simply to have the standard enable 2hr mobmod set, and added them accordingly to mob_spawn_mods.sql. Cardians and Tyger had their 2hr skills removed as they weren't necessary - Cardians got them as normal mobmods for the specific NMs, while Tyger has no mention of having Mighty Strikes on any wiki, so I commented it out for now.

This also fixes the issue of mobs being unable to use specific 2hr abilities, such as Benediction and Hundred Fists.