Scripts/SQL/Core: Implement Bar- spell magic defense bonus
Bar- spells all set the subpower to 1 for whatever reason, and a search through all of the files that contain getSubPower() had nothing in regards to bar- spells. So I'm using it to hold the bonus magic defense bonus for the element, and I removed it for the status resistance spells.

Source: https://www.bg-wiki.com/bg/Category:Barspell#Increasing_Elemental_Barspell_Potency