Updated mob pathfinding logic
Only moves when out of attack range.  Once getting in range, checks to see if another mob is close to
them.  If so, moves to their right around the target.  Should replicate the "can only fell cleave when against
a wall on your right" behaviour that retail mobs like to exhibit

Fixes #2886